% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evenFibonacci}
\alias{evenFibonacci}
\title{Even Fibonnacci}
\usage{
evenFibonacci(position, limit, total)
}
\arguments{
\item{position}{The current position}

\item{limit}{The wanted position}

\item{total}{The accumulated value}
}
\value{
The calculated result
}
\description{
\code{evenFibonnacci} sums of all of the even Fibonacci number starting at
given postion to its limits, saving in the total
}
\section{Complexity}{

This function has the following complexity \textrm{O}($n^{2}$)
}

\examples{
evenFibonacci(0, 10, 10)
evenFibonacci(5, 50, 50)
evenFibonacci(50, 1000, 870)

}
\author{
Fazendaaa
}
