% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fibonacci}
\alias{fibonacci}
\title{Seeks a number in the Fibonacci sequence}
\usage{
fibonacci(position)
}
\arguments{
\item{position}{The wanted position}
}
\description{
Seeks a number in the Fibonacci sequence
}
