% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_2.R
\name{fibonnacci}
\alias{fibonnacci}
\alias{problem2}
\title{Fibonnacci number}
\usage{
fibonnacci(limit)

problem2(limit)
}
\arguments{
\item{limit}{A positive integer.}
}
\value{
a positive integer.

The total of the even Fibonnacci numbers.
}
\description{
\code{fibonnacci} gives the Fibonnacci number of given position.
}
\section{Functions}{
\itemize{
\item \code{problem2}: \code{problem2} just filter the even Fibonnacci sequence then sum it.
}}

\examples{
fibonnacci(1) # 1
fibonnacci(2) # 2
fibonnacci(3) # 3
fibonnacci(4) # 5
problem2(1) # 1
problem2(10) # 10
problem2(40000000) # 4613732
\dontrun{
    problem2(1.2)
    problem2('a')
    problem2(c(...))
}
}
\seealso{
\url{https://projecteuler.net/problem=2} for more info about it.
}
\keyword{internal}
