% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{greatestProduct}
\alias{greatestProduct}
\title{Greatest}
\usage{
greatestProduct(numbers, digits)
}
\arguments{
\item{numbers}{Digits to iterate over}

\item{digits}{Max number of digits to compute}
}
\value{
The expected value
}
\description{
\code{greatest} finds the max number of digits multiplication
}
\section{Complexity}{

This function has the following complexity \textrm{O}($n^2$)
}

\examples{
problem8(c(1, 2, 3, 4, 5), 2)
problem8((1:10), 4)
problem8((1:100), 13)

}
\author{
Fazendaaa
}
