% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-4.R
\name{isPalindrome}
\alias{isPalindrome}
\title{Is Palindrome}
\usage{
isPalindrome(value)
}
\arguments{
\item{value}{Number to be verified}
}
\description{
Checks whether or not is a palindrome
}
\examples{
\dontrun{
isPalindrome(1)
isPalindrome(21)
isPalindrome(9009)
}
}
\keyword{internal}
