% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectEuler.R
\name{isPrime}
\alias{isPrime}
\title{Is Prime?}
\usage{
isPrime(value)
}
\arguments{
\item{value}{A integer.}
}
\value{
Boolean.
}
\description{
\code{isPrimes} checks whether or not a number is \emph{prime}
}
\examples{
isPrime(2)
isPrime(4)
isPrime(13)

}
\author{
Fazendaaa
}
