% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-4.R
\name{maxPalindrome}
\alias{maxPalindrome}
\title{Max PalindromeWhether or not is a palindrome}
\usage{
maxPalindrome(old, from, limit)
}
\arguments{
\item{old}{Max palindrome so far}

\item{from}{Lower limit to check data}

\item{limit}{Upper limit to check data}
}
\value{
The largest palindrome
}
\description{
The  largest  palindrome  from  comparssion  from  a  given  palindrome and a
sequence to be verified
}
\examples{
\dontrun{
maxPalindrome(1, 2, 9)
maxPalindrome(29, 79, 99)
maxPalindrome(126, 345, 999)
}
}
\keyword{internal}
