% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_1.r
\name{multiples}
\alias{multiples}
\alias{myFunc}
\title{Multiples of 3 or 5}
\usage{
multiples(limit)

myFunc(limit)
}
\arguments{
\item{limit}{A positive integer.}
}
\value{
A vector.

The total of the sum.
}
\description{
\code{multiples} fetch \strong{all} numbers of multiples of \emph{3} or
\emph{5} up until given limit.
}
\section{Functions}{
\itemize{
\item \code{myFunc}: \code{myFunc} just sum the values.
}}

\section{Waning}{

This is a unneeded function, it's only purpose is to make code easier to
learn how to test and document it.
}

\examples{
multiples(10) # [ 3, 5, 6, 9 ]
multiples(1000) # [ 3, 5, 6, 9, 10, ..., 999 ]
myFunc(10) # 23
myFunc(1000) # 233168
\dontrun{
myFunc(1.2)
myFunc("a")
myFunc(c(...))
}
}
\seealso{
\url{https://projecteuler.net/problem=1} for more info about it.
}
\keyword{internal}
