% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-1.R
\name{problem1}
\alias{problem1}
\title{Sum of multiples of 3 or 5}
\usage{
problem1(limit)
}
\arguments{
\item{limit}{A positive integer.}
}
\value{
The total of the sum.
}
\description{
\code{problem1} Sum all multiples of \emph{3} or \emph{5} up to given limit

\enumerate{
   \item Verifies whether or not the parameter given is compliant
   \item Do the calculations
}
}
\section{Waning}{

This is a unneeded function, it's only purpose is to make code easier to
learn how to test and document it.
}

\examples{
problem1(10)
problem1(100)
problem1(1000)

}
\seealso{
\url{https://projecteuler.net/problem=1} for more info about it.
}
\author{
Fazendaaa
}
