% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-4.R
\name{problem4}
\alias{problem4}
\title{Largest Palindrome Product}
\usage{
problem4(digits)
}
\arguments{
\item{digits}{Number of max digits to run comparison}
}
\value{
The result number
}
\description{
Calculates the largest palindrome from
}
\examples{
problem4(1)
problem4(2)
problem4(3)
}
\seealso{
\url{https://projecteuler.net/problem=4} for more info about it
}
