% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-5.R
\name{problem5}
\alias{problem5}
\title{Problem 5}
\usage{
problem5(limit)
}
\arguments{
\item{limit}{Number to be checking delimiter}
}
\value{
The smallest number
}
\description{
\code{problem5} finds the smallest positive number that is evenly divisible
by all of the numbers from 1 up to given limit
}
\section{Complexity}{

This function has the following complexity
\textrm{O}(\frac{ $n^\frac{ 5 }{ 2 }$ }{ 2 })
}

\examples{
problem5(5)
problem5(10)
problem5(20)

}
\seealso{
\url{https://projecteuler.net/problem=5} for more info about it
}
\author{
Fazendaaa
}
