% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-8.R
\name{problem8}
\alias{problem8}
\title{Largest product in a series}
\usage{
problem8(digits)
}
\arguments{
\item{digits}{Max number of digits to compute}
}
\value{
The expected value
}
\description{
\code{problem8} finds the 'n' adjacent digits in the 1000-digit number that
have the greatest product
}
\section{Complexity}{

This function has the following complexity \textrm{O}($n^2$)
}

\examples{
problem8(2)
problem8(4)
problem8(13)

}
\seealso{
\url{https://projecteuler.net/problem=8} for more info about it
}
\author{
Fazendaaa
}
