% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recursiveEvenFibonacci}
\alias{recursiveEvenFibonacci}
\title{Returns the sum of all of the even Fibonacci number starting at given
postion to its limits, saving in the total}
\usage{
recursiveEvenFibonacci(position, limit, total)
}
\arguments{
\item{position}{The current position}

\item{limit}{The wanted position}

\item{total}{The accumulated value}
}
\description{
Returns the sum of all of the even Fibonacci number starting at given
postion to its limits, saving in the total
}
