% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-4.R
\name{sequencePalindrome}
\alias{sequencePalindrome}
\title{Sequence palindrome}
\usage{
sequencePalindrome(acc, x, y)
}
\arguments{
\item{acc}{Max palindrome so far}

\item{x}{Number to be multiplied}

\item{y}{Number to be multiplied}
}
\value{
The largest palindrome
}
\description{
Multiplies two numbers, verify whehter or not this new value is a palindrome
then check whether or not is greater than the largest palindrome so far, if 
so returns it, otherwise returns the old max
}
\examples{
\dontrun{
sequencePalindrome(1, 2, 9)
sequencePalindrome(29, 79, 99)
sequencePalindrome(126, 345, 999)
}
}
\keyword{internal}
