% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-5.R
\name{smallestMultiple}
\alias{smallestMultiple}
\title{Smallest Multiple}
\usage{
smallestMultiple(limit, acc, cur)
}
\arguments{
\item{limit}{Up to tis number to be verified}

\item{acc}{The current calculus}

\item{cur}{Current value to be calculated}
}
\value{
The smallest number
}
\description{
Performs a calculus to imply the smallest multiple
}
\author{
Fazendaaa
}
\keyword{internal}
