% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectEuler.R
\name{validatePrime}
\alias{validatePrime}
\title{validatePrime}
\usage{
validatePrime(given, iterator)
}
\arguments{
\item{given}{Number to be verified}

\item{iterator}{Number to compare given to it}
}
\value{
Boolean
}
\description{
Verifies whether or not a number is prime compared to one iterated number.
}
\examples{
validatePrime(1, 1)
validatePrime(2, 4)
}
\keyword{internal}
