% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erastosthenesSieve}
\alias{erastosthenesSieve}
\title{Erastosthenes Sieve}
\usage{
erastosthenesSieve(limit)
}
\arguments{
\item{limit}{Calculation limit}
}
\value{
A list containing all prime numbers
}
\description{
\code{erastosthenesSieve} calculates the all prime numbers up till given
limit
}
\section{Complexity}{

This function has the following complexity \textrm{O}(\frac(\sqrt(n), 2))
}

\examples{
erastosthenesSieve(100)
erastosthenesSieve(200)
erastosthenesSieve(300)

}
\author{
Fazendaaa
}
