% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectEuler.R
\name{factors}
\alias{factors}
\alias{factor}
\title{Factors}
\usage{
factors(input)
}
\arguments{
\item{input}{A integer.}
}
\value{
A list of numbers.
}
\description{
Takes a number and then returns all of its factors.
}
\examples{
factors(1)
factors(123)
factors(213713)
}
