% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectEuler.R
\name{factors}
\alias{factors}
\title{Factors}
\usage{
factors(given)
}
\arguments{
\item{given}{A integer.}
}
\value{
A vector of numbers.
}
\description{
\code{factors} takes a number and then returns \strong{all} of its factors
}
\examples{
factors(1)
factors(123)
factors(456)

}
\author{
Fazendaaa
}
