% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-2.R
\name{fibonnacci}
\alias{fibonnacci}
\alias{Fibonnacci}
\alias{position}
\title{Fibonnacci number}
\usage{
fibonnacci(limit)
}
\arguments{
\item{limit}{A positive integer.}
}
\value{
a positive integer.
}
\description{
Gives the Fibonnacci number of given position.
}
\examples{
\dontrun{
fibonnacci(1)
fibonnacci(20)
}
}
\keyword{internal}
