% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectEuler.R
\name{isFactor}
\alias{isFactor}
\title{isFactor}
\usage{
isFactor(acc, cur, given)
}
\arguments{
\item{acc}{Vector with accumulated values so far}

\item{cur}{Number to be checked}

\item{given}{Number that value is or not its factor}
}
\value{
A vector of numbers that are factors or an empty vector
}
\description{
Verifies whether or not a value is a factor of another one.
}
\examples{
\dontrun{
isFactor(c(), 2, 10)
isFactor(c(2), 5, 10)
}
}
\keyword{internal}
