% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectEuler.R
\name{isPrime}
\alias{isPrime}
\title{isPrime}
\usage{
isPrime(value)
}
\arguments{
\item{value}{A integer.}
}
\value{
Boolean.
}
\description{
Checks whether or not a number is prime.
}
\examples{
isPrime(2)
isPrime(4)
}
