% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-1.R
\name{multiples}
\alias{multiples}
\title{Multiples of 3 or 5}
\usage{
multiples(limit)
}
\arguments{
\item{limit}{A positive integer.}
}
\value{
A vector.
}
\description{
\code{multiples} fetch \strong{all} numbers of multiples of \emph{3} or
\emph{5} up until given limit.
}
\examples{
\dontrun{
multiples(10)
multiples(1000)
}

}
\seealso{
\url{https://projecteuler.net/problem=1} for more info about it.
}
\author{
Fazendaaa
}
\keyword{internal}
