% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-2.R
\name{problem2}
\alias{problem2}
\alias{even}
\alias{sum}
\title{Sum of even Fibonacci numbers}
\usage{
problem2(limit)
}
\arguments{
\item{limit}{A positive integer.}
}
\value{
A total of the sum.
}
\description{
Just filter the even numbers in the Fibonnacci sequence then sum it.
}
\examples{
problem2(1)
problem2(10)
problem2(40000000)
}
\seealso{
\url{https://projecteuler.net/problem=2} for more info about it.
}
