% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-2.R
\name{problem2}
\alias{problem2}
\title{Sum of even Fibonacci numbers}
\usage{
problem2(limit)
}
\arguments{
\item{limit}{Fibonacci value limit}
}
\value{
A total of the sum.
}
\description{
\code{problem2} filter the \emph{even} numbers in the \strong{Fibonacci}
sequence then sum them.
}
\section{Complexity}{

This function has the following complexity \textrm{O}($n^{2}$)
}

\examples{
problem2(1)
problem2(10)
problem2(100)

}
\seealso{
\url{https://projecteuler.net/problem=2} for more info about it.
}
\author{
Fazendaaa
}
