% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-3.R
\name{problem3}
\alias{problem3}
\alias{largest}
\title{Largest prime factor}
\usage{
problem3(input)
}
\arguments{
\item{input}{A integer.}
}
\value{
A number
}
\description{
Largest prime factor
}
\examples{
problem3(1)
problem3(13195)
}
\seealso{
\url{https://projecteuler.net/problem=3} for more info about it.
}
