% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-3.R
\name{problem3}
\alias{problem3}
\title{Largest prime factor}
\usage{
problem3(value)
}
\arguments{
\item{value}{A integer.}
}
\value{
A number
}
\description{
\code{problem3} finds the \strong{largest} prime factor of given number
}
\section{}{
NA
}

\examples{
problem3(1)
problem3(100)
problem3(13195)

}
\seealso{
\url{https://projecteuler.net/problem=3} for more info about it.
}
\author{
Fazendaaa
}
