% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-5.R
\name{problem5}
\alias{problem5}
\title{Problem 5}
\usage{
problem5(limit)
}
\arguments{
\item{limit}{Number to be checking delimiter}
}
\value{
The smallest number
}
\description{
Problem 5
}
\seealso{
\url{https://projecteuler.net/problem=5} for more info about it
}
