% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-6.R
\name{problem6}
\alias{problem6}
\title{Sum square difference}
\usage{
problem6(limit)
}
\arguments{
\item{limit}{Limit to run the interval}
}
\value{
The expected value
}
\description{
\code{problem6} finds the difference between the sum of the squares of the
first 'n' natural numbers and the square of the sum.
}
\examples{
problem6(10)
problem6(20)
problem6(50)

}
\seealso{
\url{https://projecteuler.net/problem=6} for more info about it
}
\author{
Fazendaaa
}
