% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem-7.R
\name{problem7}
\alias{problem7}
\title{Nth prime}
\usage{
problem7(limit)
}
\arguments{
\item{limit}{Limit to run the interval}
}
\value{
The expected value
}
\description{
\code{problem7} finds the nth prime
}
\section{Complexity}{

This function has the following complexity \textrm{O}(n$^\frac(3, 2)$)
}

\examples{
problem7(13)
problem7(50)
problem7(100)

}
\seealso{
\url{https://projecteuler.net/problem=7} for more info about it
}
\author{
Fazendaaa
}
